Remodularize client and config
