+++ gneiss-mqtt/build-matrix.sh
@@ -4,18 +4,16 @@ set -e
 
 cargo clean
 echo "Building with no features"
-cargo build
+cargo build --features=strict
 
 declare -a features=("threaded" "threaded-rustls" "threaded-native-tls" "threaded-websockets" "threaded-websockets,threaded-rustls" "threaded-websockets,threaded-native-tls" "tokio" "tokio-rustls" "tokio-native-tls" "tokio-websockets" "tokio-websockets,tokio-rustls" "tokio-websockets,tokio-native-tls")
 
 for i in "${features[@]}"
 do
-   cargo clean
    echo "Building with features: $i"
-   cargo build --features=$i
+   cargo build --features=strict,$i
 done
 
-cargo clean
 echo "Building with all features"
 cargo build --all-features
