+++ gneiss-mqtt/src/client/config.rs
@@ -9,8 +9,6 @@ Module containing types for configuring an MQTT client.
 
 use crate::alias::{OutboundAliasResolverFactoryFn};
 use crate::error::*;
-#[cfg(any(feature="tokio", feature="threaded"))]
-use crate::client::*;
 use crate::mqtt::*;
 
 use log::*;
@@ -30,12 +28,14 @@ use std::{future::Future, pin::Pin,};
 use tungstenite::{client::*, handshake::client::generate_key};
 
 #[cfg(feature="tokio")]
-use crate::features::tokio::*;
+use crate::client::asynchronous::*;
 #[cfg(feature="tokio")]
-use tokio::runtime::Handle;
+use crate::client::asynchronous::tokio::{TokioClientOptions, make_client_tokio};
 
 #[cfg(feature="threaded")]
-use crate::features::threaded::*;
+use crate::client::synchronous::*;
+#[cfg(feature="threaded")]
+use crate::client::synchronous::threaded::*;
 
 /// Configuration options related to establishing connections through HTTP proxies
 #[derive(Default, Clone)]
@@ -781,171 +781,6 @@ impl MqttClientOptionsBuilder {
     /// Builds a new set of client options
     pub fn build(&self) -> MqttClientOptions {
         self.options.clone()
-    }
-}
-
-/// A structure that holds configuration related to a client's asynchronous properties and
-/// internal implementation.  Only relevant to asynchronous clients.
-pub struct AsyncClientOptions {
-
-    #[cfg(feature="tokio-websockets")]
-    pub(crate) websocket_options: Option<AsyncWebsocketOptions>
-}
-
-/// Builder type for asynchronous client behavior.
-pub struct AsyncClientOptionsBuilder {
-    options: AsyncClientOptions
-}
-
-impl AsyncClientOptionsBuilder {
-
-    /// Creates a new builder object for AsyncClientOptions
-    pub fn new() -> Self {
-        AsyncClientOptionsBuilder {
-            options: AsyncClientOptions {
-                #[cfg(feature="tokio-websockets")]
-                websocket_options: None
-            }
-        }
-    }
-
-    #[cfg(feature="tokio-websockets")]
-    /// Configures an asynchronous client to use websockets for MQTT transport
-    pub fn with_websocket_options(&mut self, websocket_options: AsyncWebsocketOptions) -> &mut Self {
-        self.options.websocket_options = Some(websocket_options);
-        self
-    }
-
-    /// Builds a new set of asynchronous client options
-    pub fn build(self) -> AsyncClientOptions {
-        self.options
-    }
-}
-
-impl Default for AsyncClientOptionsBuilder {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-#[cfg(feature="tokio")]
-/// A structure that holds configuration related to how an asynchronous client should interact
-/// with the Tokio async runtime.
-pub struct TokioClientOptions {
-    pub(crate) runtime: Handle,
-}
-
-#[cfg(feature="tokio")]
-/// Builder type for tokio-based client configuration
-pub struct TokioClientOptionsBuilder {
-    options: TokioClientOptions
-}
-
-#[cfg(feature="tokio")]
-impl TokioClientOptionsBuilder {
-
-    /// Creates a new builder object for TokioClientOptions
-    pub fn new(runtime: Handle) -> Self {
-        TokioClientOptionsBuilder {
-            options: TokioClientOptions {
-                runtime
-            }
-        }
-    }
-
-    /// Builds a new set of tokio client configuration options
-    pub fn build(self) -> TokioClientOptions {
-        self.options
-    }
-}
-
-/// A structure that holds configuration related to a client's synchronous properties and
-/// internal implementation.  Only relevant to synchronous clients.
-pub struct SyncClientOptions {
-    #[cfg(feature="threaded-websockets")]
-    pub(crate) websocket_options: Option<SyncWebsocketOptions>
-}
-
-/// Builder type for synchronous client behavior.
-pub struct SyncClientOptionsBuilder {
-    config: SyncClientOptions
-}
-
-impl SyncClientOptionsBuilder {
-
-    /// Creates a new builder object for SyncClientOptions
-    pub fn new() -> Self {
-        SyncClientOptionsBuilder {
-            config: SyncClientOptions {
-                #[cfg(feature="threaded-websockets")]
-                websocket_options: None,
-            }
-        }
-    }
-
-    #[cfg(feature="threaded-websockets")]
-    /// Configures a synchronous client to use websockets for MQTT transport
-    pub fn with_websocket_options(&mut self, websocket_options: SyncWebsocketOptions) -> &mut Self {
-        self.config.websocket_options = Some(websocket_options);
-        self
-    }
-
-    /// Builds a new set of synchronous client options
-    pub fn build(self) -> SyncClientOptions {
-        self.config
-    }
-}
-
-impl Default for SyncClientOptionsBuilder {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-#[cfg(feature="threaded")]
-/// Threaded client specific configuration
-pub struct ThreadedClientOptions {
-    pub(crate) idle_service_sleep: Option<Duration>,
-}
-
-#[cfg(feature="threaded")]
-/// Builder type for threaded client configuration
-pub struct ThreadedClientOptionsBuilder {
-    config: ThreadedClientOptions
-}
-
-#[cfg(feature="threaded")]
-impl ThreadedClientOptionsBuilder {
-
-    /// Creates a new builder object for ThreadedClientOptions
-    pub fn new() -> Self {
-        ThreadedClientOptionsBuilder {
-            config: ThreadedClientOptions {
-                idle_service_sleep: None,
-            }
-        }
-    }
-
-    /// Configures the time interval to sleep the thread the client runs on between io
-    /// processing events.  Only used if no events occurred on the previous iteration.  If the
-    /// client is handling significant work, it will not sleep, but if there's nothing
-    /// happening, it will.
-    ///
-    /// If not set, defaults to 20 milliseconds.
-    pub fn with_idle_service_sleep(&mut self, duration: Duration) {
-        self.config.idle_service_sleep = Some(duration);
-    }
-
-    /// Builds a new set of threaded client configuration options
-    pub fn build(self) -> ThreadedClientOptions {
-        self.config
-    }
-}
-
-#[cfg(feature="threaded")]
-impl Default for ThreadedClientOptionsBuilder {
-    fn default() -> Self {
-        Self::new()
     }
 }
